html/canvas/offscreen/drawing-images-to-the-canvas/2d.drawImage.broken.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.drawImage.broken</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.drawImage.broken</h1>
<p class="desc"></p>


<script>
promise_test(async t => {

var canvas = new OffscreenCanvas(100, 50);
var ctx = canvas.getContext('2d');

const response = await fetch('/images/broken.png');
const blob = await response.blob();

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
await promise_rejects_dom(t, 'InvalidStateError', createImageBitmap(blob), 'The source image could not be decoded.');
t.done();

}, "");
</script>
